Move the code to discord-api file
